🐛 improve documentation, support multiple identical slugs
